LayoutTests/fast/workers/worker-crash-with-invalid-location-expected.txt

Blocked access to external URL http://example.com/worker.js
CONSOLE MESSAGE: Cannot load http://example.com/worker.js due to access control checks.
CONSOLE MESSAGE: Worker failed to load.
Test worker fetch of blocked url. Should print a "PASS" statement.

PASS
